fix mystakes
