+++ src/main.js
@@ -3,7 +3,7 @@ import {createStatisticTemplate} from "./components/statistic";
 import {createFilterTemplate} from "./components/filter";
 import {createFilmsTemplate} from "./components/films";
 import {createFilmCardTemplate} from "./components/film-card";
-import {createFilmDetales} from "./components/film-detales";
+import {createFilmDetales} from "./components/film-details";
 import {createShowMoreButton} from "./components/show-more-btn";
 import {extraFilmsTemplate} from "./components/extra-films-template";
 import {generateCards} from "./mock/card";
@@ -68,54 +68,53 @@ showMoreButton.addEventListener(`click`, () => {
 });
 
 const extraFilmElement = siteMainElement.querySelector(`.films`);
-render(extraFilmElement, extraFilmsTemplate(`Top Rated`, `top-rated`));
-render(extraFilmElement, extraFilmsTemplate(`Most Commented`, `most-commented`));
+render(extraFilmElement, extraFilmsTemplate(`Top Rated`));
+render(extraFilmElement, extraFilmsTemplate(`Most Commented`));
 
 
-const filmContainerElement = [...extraFilmElement.querySelectorAll(`.films-list--extra`)];
+// const filmContainerElement = [...extraFilmElement.querySelectorAll(`.films-list--extra`)];
 
 const getTopRatedFilms = (cardsArray) => {
-  const sortedCards = cardsArray.slice().sort((a, b) => {
-    if (a.rating > b.rating) {
-      return 1;
-    }
-    if (a.rating < b.rating) {
-      return -1;
-    }
-    return 0;
-  });
+  const sortedCards = cardsArray.slice().sort((a, b) => a.rating - b.rating);
   sortedCards.splice(0, sortedCards.length - 2);
   return sortedCards;
 };
 
 const getTopCommentsFilms = (cardsArray) => {
-  const sortedCards = cardsArray.slice().sort((a, b) => {
-    if (a.comments > b.comments) {
-      return 1;
-    }
-    if (a.comments < b.comments) {
-      return -1;
-    }
-    return 0;
-  });
+  const sortedCards = cardsArray.slice().sort((a, b) => a.comments.length - b.comments.length);
   sortedCards.splice(0, sortedCards.length - 2);
   return sortedCards;
 };
 
-const tail = (list) => list[list.length - 1];
+const top = [
+  getTopRatedFilms(cards),
+  getTopCommentsFilms(cards)
+];
 
-const topRatedElement = filmContainerElement[0].querySelector(`.films-list__container`);
-const topRatedCards = getTopRatedFilms(cards);
-topRatedCards.forEach((card) => {
-  render(topRatedElement, createFilmCardTemplate(card));
-  const cardElement = tail(topRatedElement.querySelectorAll(`.film-card`));
-  cardElement.addEventListener(`click`, () => handleFilmCardClick(card));
+const filmContainerElement = extraFilmElement.querySelectorAll(`.films-list--extra .films-list__container`);
+
+filmContainerElement.forEach((item, index) => {
+  top[index].forEach((card) => {
+    render(item, createFilmCardTemplate(card));
+    const cardElement = item.querySelector(`.film-card:last-child`);
+    cardElement.addEventListener(`click`, () => handleFilmCardClick(card));
+  });
 });
 
-const topCommentedElement = filmContainerElement[1].querySelector(`.films-list__container`);
-const topCommentsCards = getTopCommentsFilms(cards);
-topCommentsCards.forEach((card) => {
-  render(topCommentedElement, createFilmCardTemplate(card));
-  const cardElement = tail(topCommentedElement.querySelectorAll(`.film-card`));
-  cardElement.addEventListener(`click`, () => handleFilmCardClick(card));
-});
+// const tail = (list) => list[list.length - 1];
+
+// const topRatedElement = filmContainerElement[0].querySelector(`.films-list__container`);
+// const topRatedCards = getTopRatedFilms(cards);
+// topRatedCards.forEach((card) => {
+//   render(topRatedElement, createFilmCardTemplate(card));
+//   const cardElement = tail(topRatedElement.querySelectorAll(`.film-card`));
+//   cardElement.addEventListener(`click`, () => handleFilmCardClick(card));
+// });
+
+// const topCommentedElement = filmContainerElement[1].querySelector(`.films-list__container`);
+// const topCommentsCards = getTopCommentsFilms(cards);
+// topCommentsCards.forEach((card) => {
+//   render(topCommentedElement, createFilmCardTemplate(card));
+//   const cardElement = tail(topCommentedElement.querySelectorAll(`.film-card`));
+//   cardElement.addEventListener(`click`, () => handleFilmCardClick(card));
+// });
